+++ src/components/layout/header/header-link.tsx
@@ -2,13 +2,13 @@ import Link from 'next/link';
 import { ReactNode } from 'react';
 
 export interface IHeaderLink {
-  href: string;
-  content: string | ReactNode;
+  href?: string;
+  content?: string | ReactNode;
 }
 
 export const HeaderLink = ({ href, content }: IHeaderLink) => (
-  <Link href={href}>
-    <span className="p-4 h-full hover:bg-gray-600 hover:text-white transition duration-200 cursor-pointer">
+  <Link href={href || '/'}>
+    <span className="hover:text-gray-500 hover:underline text-gray-600 flex flex-row justify-center items-center p-1 cursor-pointer font-mono">
       {content}
     </span>
   </Link>