+++ src/pages/index.tsx
@@ -1,10 +1,9 @@
 import { NextPage } from 'next';
-import Link from 'next/link';
 
 const Home: NextPage = () => {
   return (
     <div className="w-full flex flex-col space-y-4 justify-center items-center">
-      <Link href="/implementation-references">Implementation References</Link>
+      <h1>Navegue pelas páginas através do cabeçalho</h1>
     </div>
   );
 };