+++ next.config.js
@@ -1,4 +1,5 @@
 /** @type {import('next').NextConfig} */
+require('reflect-metadata')
 module.exports = {
   reactStrictMode: true,
 }