+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 2021.11.16
 
 - `yarn create next-app --typescript`: bootstrap nextjs project
+- `yarn add -D prettier`: add prettier