+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - `yarn create next-app --typescript`: bootstrap nextjs project
 - `yarn add -D prettier`: add prettier
 - add editorconfig
+- fix default project structure by arranging `pages` and `styles` folders inside `src`