+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - `yarn add -D prettier`: add prettier
 - add editorconfig
 - fix default project structure by arranging `pages` and `styles` folders inside `src`
+- [tailwindcss configuration](https://tailwindcss.com/docs/guides/nextjs)